--- a/p5js/coding-challenges/marching-squares-2/README.md
+++ b/p5js/coding-challenges/marching-squares-2/README.md
@@ -1,0 +1,36 @@
+
+## Overview
+
+This is a [p5.js][p5js-home] sketch implementing the basic Marching Squares algorithm, inspired by [Coding Train's video - Coding Marching Squares
+][coding-train-yt-marching-sq].
+
+Next steps in upcoming sketches / iterations
+* Fix the issue of not connecting the isolines to the left and top window edges.
+* Enhance this by using interpolation to have smoother lines
+* Allow for multiple steps within the 2D value field
+
+
+## Controls
+
+(No interactivity implemented yet).
+
+# References:
+* [Marching Squares (Wikipedia)] article describing the algorithm
+* [Coding Train's video - Marching Squares][coding-train-yt-marching-sq] - original inspiring video.
+
+# Links:
+
+* [Live View][live-view]
+* [Source on Github][source-code]
+
+# Screenshot:
+
+![screenshot][screenshot-01]
+
+[p5js-home]: https://p5js.org/
+[source-code]: https://github.com/brianhonohan/sketchbook/tree/master/p5js/coding-challenges/marching-squares/
+[live-view]: https://brianhonohan.com/sketchbook/p5js/coding-challenges/marching-squares/
+[screenshot-01]: ./screenshot-01.png
+
+[coding-train-yt-marching-sq]: https://www.youtube.com/watch?v=0ZONMNUKTfU
+[wikipedia-marching-squares]: https://en.wikipedia.org/wiki/Marching_squares