--- a/p5js/coding-challenges/marching-squares-2/app.js
+++ b/p5js/coding-challenges/marching-squares-2/app.js
@@ -31,6 +31,7 @@ function setup() {
   // P5JsSettings.addObject(gui, system, regenerateSystem);
 
   // Option 3: method chaining
+  gui.add(system.settings, "cellWidth", 1, 100, 1).onChange(reinitSystem);
   gui.add(system.settings, "scale", 0.01, 0.2, 0.001).onChange(regenerateSystem);
   gui.add(system.settings, "xOffset", -10000, 10000, 1).onChange(regenerateSystem);
   gui.add(system.settings, "yOffset", -10000, 10000, 1).onChange(regenerateSystem);
@@ -41,6 +42,10 @@ function setup() {
 
 function regenerateSystem(){
   system.regenerate();
+}
+
+function reinitSystem(){
+  system.init();
 }
 
 function draw(){