--- a/p5js/coding-challenges/marching-squares-2/index.md
+++ b/p5js/coding-challenges/marching-squares-2/index.md
@@ -1,0 +1,27 @@
+---
+layout: minimal
+title:  "p5.js - Marching Squares 2"
+categories: p5js
+modal: true
+viewport_noscale: true
+excerpt: An iteration of the 'Marching Squares' to add in interpolation and multi-level isolines.
+
+js_scripts:
+- /sketchbook/vendor/inobounce/0.2.0/inobounce.js
+- /sketchbook/vendor/p5.js/1.11.2/p5.min.js
+- /sketchbook/vendor/dat.gui/0.7.7/dat.gui.js
+- /sketchbook/p5js/common/p5js_settings.js
+- /sketchbook/p5js/common/p5js_utils.js
+- /sketchbook/js/util_functions.js
+- /sketchbook/js/options_set.js
+- /sketchbook/js/models/rect.js
+- /sketchbook/js/models/cell_grid.js
+- classes/cell.js
+- classes/cell_viewer.js
+- classes/system.js
+- app.js
+
+---
+
+{% include_relative README.md %}
+