--- a/p5js/coding-challenges/marching-squares-2/local_dev.html
+++ b/p5js/coding-challenges/marching-squares-2/local_dev.html
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta content="text/html;charset=utf-8" http-equiv="Content-Type">
+    <meta content="utf-8" http-equiv="encoding">
+    <title>P5.JS - Marching Squares 2</title>
+    <meta name="viewport" content="user-scalable=no,initial-scale=1,maximum-scale=1,minimum-scale=1,width=device-width">
+
+    <script src="/sketchbook/vendor/inobounce/0.2.0/inobounce.js" charset="utf-8"></script>
+    <script src="/sketchbook/vendor/p5.js/1.11.2/p5.min.js"></script>
+    <script src="/sketchbook/vendor/dat.gui/0.7.7/dat.gui.js"></script>
+    <script src="/sketchbook/p5js/common/p5js_settings.js"></script>
+    <script src="/sketchbook/p5js/common/p5js_utils.js"></script>
+    <script src="/sketchbook/js/util_functions.js"></script>
+    <script src="/sketchbook/js/options_set.js"></script>
+    <script src="/sketchbook/js/models/rect.js"></script>
+    <script src="/sketchbook/js/models/cell_grid.js"></script>
+
+    <script src="classes/cell.js" charset="utf-8"></script>
+    <script src="classes/cell_viewer.js" charset="utf-8"></script>
+    <script src="classes/system.js" charset="utf-8"></script>
+    <script src="app.js" charset="utf-8"></script>
+
+<style>
+body {
+  margin: 0;
+  font-family: Sans-Serif;
+}
+</style>
+</head>
+<body>
+</body>
+</html>