… p5js/coding-challenges/marching-squares-2/screenshot-01.png