Still a work in progress, and will refine this sketch.

But this makes a couple of improvements

* Rendering directly in a simple loop; 
    - rather than a loop calling a render function; within which I was setting styling like stroke() fill() repeatedly.
* Simpler way to add listeners to dat.gui config

----

Rendering optimizations

### Before

Cell Width: 15px: 8 ms / frame
 
![01_900x900_cw15-Screenshot 2025-01-25 182400](https://github.com/user-attachments/assets/16c1d938-e687-4518-aa5b-f36eb3857cf1)

Cell Width 12px 65 ms / frame
- starting to see frame drops

![01_900x900-cw12-Screenshot 2025-01-25 182702](https://github.com/user-attachments/assets/3e432a9c-78e8-460f-82ba-1fc6f8c32357)


### After

Cell Width: 15px: 3.2 ms / frame

![02_900x900_cw15-Screenshot 2025-01-25 183200](https://github.com/user-attachments/assets/3437341b-a3ae-484f-9bf7-6796bd0aa703)

Cell width 12 px 4.5 ms / frame

![02-900x900_cw12-Screenshot 2025-01-25 183238](https://github.com/user-attachments/assets/abf9a34a-52a9-4a06-8c22-af88925acce1)

Cell width 8px : 9ms / frame

![02_900x900_cw8-Screenshot 2025-01-25 183323](https://github.com/user-attachments/assets/ce6aefb5-8fa5-4f6b-bdb8-f0b0d0b3b734)